PIX-2411 better storage organization
