+++ src/main/java/com/appmetr/s2s/persister/BatchStorage.java
@@ -33,5 +33,5 @@ public interface BatchStorage {
     /**
      * Removes the head of this storage.
      */
-    void remove() throws IOException;
+    void remove();
 }
